pyproject.toml
[project]
name = "audio-metrics"
version = "0.1.0"
description = "Metrics to measure the quality of audio"
authors = [{name = "Maarten Grachten"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"torch<2.6",
	"torchvision",
	"laion-clap @ git+https://github.com/LAION-AI/CLAP.git",
	"soundfile",
	"pyloudnorm",
	"scikit-learn",
	"loguru",
	"cylimiter",
	"prdc",
	"einops",
	"resampy",
	"tqdm",
	"rich",
	"appdirs",
]

[tool.uv]
package = true

[tool.black]
line-length = 90

[tool.ruff]
line-length = 90

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=8.34.0",
    "opt-einsum>=3.4.0",
    "python-benedict>=0.34.1",
    "python-lsp-server>=1.12.2",
    "ruff>=0.11.1",
]

